Update for sql script of shopping mall, and the Bless of the Buddha.

+++ src/main/java/com/tia102g5/bookticket/model/BookTicket.java
@@ -1,6 +1,7 @@
 package com.tia102g5.bookticket.model;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 import java.util.Set;
 
@@ -56,7 +57,7 @@ public class BookTicket implements Serializable {
 	private Integer ticketQuantity; // 數量
 
 	@Column(name = "totalPrice")
-	private Integer totalPrice; // 總金額
+	private BigDecimal totalPrice; // 總金額
 	
 	@OneToMany(mappedBy = "bookTicket", cascade = CascadeType.ALL)
 	@OrderBy("ticketID asc")
@@ -69,7 +70,7 @@ public class BookTicket implements Serializable {
 
 	public BookTicket(Integer bookTicketID, GeneralMember generalMember, Activity activity,
 			ActivityTimeSlot activityTimeSlot, MemberCoupon memberCoupon, Timestamp bookTime, Integer ticketQuantity,
-			Integer totalPrice, Set<Ticket> tickets) {
+			BigDecimal totalPrice, Set<Ticket> tickets) {
 		super();
 		this.bookTicketID = bookTicketID;
 		this.generalMember = generalMember;
@@ -139,11 +140,11 @@ public class BookTicket implements Serializable {
 		this.ticketQuantity = ticketQuantity;
 	}
 
-	public Integer getTotalPrice() {
+	public BigDecimal getTotalPrice() {
 		return totalPrice;
 	}
 
-	public void setTotalPrice(Integer totalPrice) {
+	public void setTotalPrice(BigDecimal totalPrice) {
 		this.totalPrice = totalPrice;
 	}
